UpdateForConflictBetweenManagers

+++ ITI.CEI40.Monitor/Models/View Models/EngineerChrtViewModel.cs
@@ -10,9 +10,9 @@ namespace ITI.CEI40.Monitor.Models.View_Models
     public class EngineerChrtViewModel
     {
         public List<string> Months { get; set; }
-        public List<int> Quality { get; set; }
-        public List<int> Time { get; set; }
-        public List<int> Complexity { get; set; }
+        public List<float> Quality { get; set; }
+        public List<float> Time { get; set; }
+        public List<float> Complexity { get; set; }
         public List<SubTask> LastTasks { get; set; }
         public string EngineerName { get; set; }
